requirements.txt
fastapi==0.115.12
joblib==1.4.2
numpy==2.0.2
pandas==2.2.3
pydantic==2.11.3
python-dotenv==1.1.0
scikit-learn==1.6.1
scipy==1.13.1
uvicorn==0.34.0
sqlalchemy[asyncio]  
psycopg2-binary>=2.9
databases  
requests
supabase
streamlit
